requirements.txt
streamlit
numpy
pandas
opencv-python-headless
astar
Pillow
simpleai
tk
streamlit-drawable-canvas
